oops wanted to add in different commits, sorry. cut fat.
