+++ Assets/Enemy.prefab
@@ -29,7 +29,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 6736919064825838575}
-  m_RootOrder: 2
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!58 &1213223814
 CircleCollider2D:
@@ -75,8 +75,90 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 6736919064825838575}
-  m_RootOrder: 1
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &4118924303705989941
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 6907585736078068475}
+  - component: {fileID: 8500832485977277530}
+  m_Layer: 0
+  m_Name: base
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &6907585736078068475
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4118924303705989941}
+  m_LocalRotation: {x: -0, y: -0, z: 0.9986424, w: 0.05209085}
+  m_LocalPosition: {x: 0.026732706, y: -0.010324076, z: 0}
+  m_LocalScale: {x: 0.40558144, y: 0.40558144, z: 0.40558144}
+  m_Children: []
+  m_Father: {fileID: 6736919064825838575}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 174.028}
+--- !u!212 &8500832485977277530
+SpriteRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4118924303705989941}
+  m_Enabled: 1
+  m_CastShadows: 0
+  m_ReceiveShadows: 0
+  m_DynamicOccludee: 1
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
+  m_RayTraceProcedural: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 0
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_Sprite: {fileID: 21300000, guid: 36a6c56df6ae64b078e8331130533d7f, type: 3}
+  m_Color: {r: 1, g: 0, b: 0, a: 1}
+  m_FlipX: 0
+  m_FlipY: 0
+  m_DrawMode: 0
+  m_Size: {x: 4, y: 4}
+  m_AdaptiveModeThreshold: 0.5
+  m_SpriteTileMode: 0
+  m_WasSpriteAssigned: 1
+  m_MaskInteraction: 0
+  m_SpriteSortPoint: 0
 --- !u!1 &6736919064825838547
 GameObject:
   m_ObjectHideFlags: 0
@@ -89,7 +171,6 @@ GameObject:
   - component: {fileID: 6736919064825838574}
   - component: {fileID: 6736919064825838544}
   - component: {fileID: 6736919064825838546}
-  - component: {fileID: 9028507426979333424}
   - component: {fileID: 4225439136666806338}
   - component: {fileID: 4225439136666806349}
   - component: {fileID: 4225439136666806336}
@@ -111,9 +192,9 @@ Transform:
   m_LocalPosition: {x: 0.05000019, y: 0.16, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
-  - {fileID: 3579995429934525593}
   - {fileID: 7876153354087013304}
   - {fileID: 1213223813}
+  - {fileID: 6907585736078068475}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -126,12 +207,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 6736919064825838547}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a49fc6cbcd9d0c94cb2e00a7a397b84c, type: 3}
+  m_Script: {fileID: 11500000, guid: e9aec5339ca304d9ea61dbfd561cae1c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  pitfallAnimSpeed: 0.015
-  respawnDistFromPit: 2
-  customRespawnLocation: {fileID: 0}
+  speed: 4
+  dest: {fileID: 0}
 --- !u!212 &6736919064825838544
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -180,7 +260,7 @@ SpriteRenderer:
   m_Size: {x: 0.2, y: 0.2}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
-  m_WasSpriteAssigned: 1
+  m_WasSpriteAssigned: 0
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
 --- !u!50 &6736919064825838546
@@ -204,20 +284,6 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 4
---- !u!114 &9028507426979333424
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6736919064825838547}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 18372fd4eb7fc46dd9e663220936efc5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 4
-  dest: {fileID: 0}
 --- !u!61 &4225439136666806338
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -300,65 +366,3 @@ MonoBehaviour:
     serializedVersion: 2
     rgba: 4292990208
   path: []
---- !u!1001 &6736919064740320519
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 6736919064825838575}
-    m_Modifications:
-    - target: {fileID: 7841025139664992669, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_Name
-      value: PitfallCollider
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -0.39
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
---- !u!4 &3579995429934525593 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 7841025139664992670, guid: be5c5d3a3f4e3db4e98588ac64ebfb9f, type: 3}
-  m_PrefabInstance: {fileID: 6736919064740320519}
-  m_PrefabAsset: {fileID: 0}
