+++ Assets/Scenes/Main.unity
@@ -2746,7 +2746,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &899522970
 Transform:
   m_ObjectHideFlags: 0
@@ -4742,8 +4742,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0ea194af97b1041558c003953e4e143e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  e: {fileID: 9028507426979333424, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-  isPath: 0
+  e: {fileID: 6736919064825838574, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
+  isPath: 1
   target: {fileID: 409039200}
   uicon: {fileID: 1725121961}
 --- !u!4 &1031168517
@@ -4756,8 +4756,7 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -7.15, y: -0.14, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children:
-  - {fileID: 7484312650118908955}
+  m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -5565,77 +5564,3 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2052505593}
   m_CullTransparentMesh: 1
---- !u!1001 &4225439135873955828
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 1031168517}
-    m_Modifications:
-    - target: {fileID: 4225439136666806336, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: es
-      value: 
-      objectReference: {fileID: 1031168516}
-    - target: {fileID: 4225439136666806336, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: Target
-      value: 
-      objectReference: {fileID: 409039200}
-    - target: {fileID: 6736919064825838547, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_Name
-      value: Enemy
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0.05000019
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0.16
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9028507426979333424, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-      propertyPath: dest
-      value: 
-      objectReference: {fileID: 409039196}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
---- !u!4 &7484312650118908955 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 6736919064825838575, guid: 8a7ed305039ed40a2a095e518a4396c0, type: 3}
-  m_PrefabInstance: {fileID: 4225439135873955828}
-  m_PrefabAsset: {fileID: 0}
