--- Assets/Topdown2DPitfall.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ef9eaa8eb10f340a2a8553f8ad7d76c2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
