--- Assets/Topdown2DPitfall/DemoScene.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5c52a040ebc2a684ab946f7d90336444
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
