--- Assets/Topdown2DPitfall/DemoScene/DemoArtAndTiles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 328791b6b5457164e930f7a3a5f2d4a7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
