--- Assets/Topdown2DPitfall/DemoScene/DemoArtAndTiles/floor_1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c629b1b12b6dbc04e8d6456c4ee4cef2
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
