--- Assets/Topdown2DPitfall/DemoScene/DemoArtAndTiles/roguelikeSheet_transparent_0.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8e0c3f18223a37c43b5d08f5c08de192
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
