--- Assets/Topdown2DPitfall/DemoScene/DemoPrefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 754384a49e5077f41b63f87f42996aeb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
