--- Assets/Topdown2DPitfall/DemoScene/DemoPrefabs/DemoTilePalette.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4aee82597cc7c6c429febdea05119830
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
