--- Assets/Topdown2DPitfall/DemoScene/DemoPrefabs/Enemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2ebf8ace48898435cad857664fe03d28
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
