--- Assets/Topdown2DPitfall/DemoScene/DemoPrefabs/Player.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 47f9d147ca95a534f834556c9efe8694
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
