--- Assets/Topdown2DPitfall/DemoScene/DemoPrefabs/Tilemap.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: baf51a99397ec3d42a78e01938ea6a46
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
