--- Assets/Topdown2DPitfall/DemoScene/DemoScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c2c13863f394fef45b1d0199ecd98bb7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
