--- Assets/Topdown2DPitfall/Documentation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8c80e4824beee5146a0a15543031f8f0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
