--- Assets/Topdown2DPitfall/Documentation/Topdown 2D Pitfall System.pdf.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 29b3d88535eeedb4fafb991550a9537e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
