--- Assets/Topdown2DPitfall/PitfallCollider.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: be5c5d3a3f4e3db4e98588ac64ebfb9f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
