--- Assets/Topdown2DPitfall/PitfallManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d8fc1eba0016dc649aafb88a6d56a928
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
