--- Assets/Topdown2DPitfall/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8678677746e5c584ab544323baa2fb92
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
