--- Assets/Topdown2DPitfall/Scripts/Interfaces.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: be025e4b6e9e97a4f9cb1a50b4944fb3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
