--- Assets/Topdown2DPitfall/Scripts/Interfaces/IPitfall.cs
@@ -1,5 +1,0 @@
-﻿namespace Nevelson.Topdown2DPitfall.Assets.Scripts.Utils {
-    public interface IPitfall {
-        void TriggerPitfall();
-    }
-}
