--- Assets/Topdown2DPitfall/Scripts/Interfaces/IPitfall.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd042c073e701b14daab280c59976f6c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
