--- Assets/Topdown2DPitfall/Scripts/Interfaces/IPitfallCheck.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6fe19dc082b372f41a6cb9a05c1a4b85
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
