--- Assets/Topdown2DPitfall/Scripts/Interfaces/IPitfallObject.cs
@@ -1,6 +1,0 @@
-﻿namespace Nevelson.Topdown2DPitfall.Assets.Scripts.Utils {
-    public interface IPitfallObject {
-        void PitfallActionsBefore();
-        void PitfallResultingAfter();
-    }
-}
