--- Assets/Topdown2DPitfall/Scripts/Pitfall.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9a4d2cf0ce767947bb5ab8c00ab6022
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
