--- Assets/Topdown2DPitfall/Scripts/PitfallColliderLayerSetter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d66e2ab4284f5d74abfa6c370ef024e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
