--- Assets/Topdown2DPitfall/Scripts/PitfallObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a49fc6cbcd9d0c94cb2e00a7a397b84c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
