--- Assets/Topdown2DPitfall/Scripts/Utils.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1aab538aae2ea0c4cb32d683b3b69d91
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
