--- Assets/Topdown2DPitfall/Scripts/Utils/Constants.cs
@@ -1,5 +1,0 @@
-﻿namespace Nevelson.Topdown2DPitfall.Assets.Scripts.Utils {
-    public static class Constants {
-        public const string PITFALL_COLLIDER = "PitfallCollider";
-    }
-}
