--- Assets/Topdown2DPitfall/Scripts/Utils/ExtTransform.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 217913f249e10ba478de1d90df72ef69
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
